Pulled backgroundfecthtask from backgroundtask service
